+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## v1.7.2 - 2024-10-22
+
+### Fixed
+
+- calculation of changed cron patterns
+
 ## v1.7.1 - 2024-10-22
 
 ### Fixed